Renamed to AndroidPlatform
