+++ android/jni/app/organicmaps/MwmApplication.cpp
@@ -1,7 +1,7 @@
 #include "app/organicmaps/Framework.hpp"
 
 #include "app/organicmaps/platform/GuiThread.hpp"
-#include "app/organicmaps/platform/PlatformNew.hpp"
+#include "app/organicmaps/platform/AndroidPlatform.hpp"
 
 #include "app/organicmaps/core/jni_helper.hpp"
