+++ android/jni/app/organicmaps/TrafficState.cpp
@@ -2,7 +2,7 @@
 
 #include "app/organicmaps/core/jni_helper.hpp"
 
-#include "app/organicmaps/platform/PlatformNew.hpp"
+#include "app/organicmaps/platform/AndroidPlatform.hpp"
 
 extern "C"
 {
