+++ android/jni/app/organicmaps/editor/OpeningHours.cpp
@@ -1,7 +1,7 @@
 #include <jni.h>
 
 #include "app/organicmaps/core/jni_helper.hpp"
-#include "app/organicmaps/platform/PlatformNew.hpp"
+#include "app/organicmaps/platform/AndroidPlatform.hpp"
 
 #include "editor/opening_hours_ui.hpp"
