+++ android/jni/app/organicmaps/opengl/androidoglcontextfactory.cpp
@@ -1,7 +1,7 @@
 #include "androidoglcontextfactory.hpp"
 #include "android_gl_utils.hpp"
 
-#include "app/organicmaps/platform/PlatformNew.hpp"
+#include "app/organicmaps/platform/AndroidPlatform.hpp"
 
 #include "base/assert.hpp"
 #include "base/logging.hpp"
