+++ android/jni/app/organicmaps/platform/AndroidPlatform.hpp
