+++ android/jni/app/organicmaps/platform/Localization.cpp
@@ -2,7 +2,7 @@
 
 #include "app/organicmaps/core/jni_helper.hpp"
 #include "app/organicmaps/core/ScopedLocalRef.hpp"
-#include "app/organicmaps/platform/PlatformNew.hpp"
+#include "app/organicmaps/platform/AndroidPlatform.hpp"
 
 #include "platform/localization.hpp"
