+++ android/jni/app/organicmaps/vulkan/android_vulkan_context_factory.cpp
@@ -1,6 +1,6 @@
 #include "android_vulkan_context_factory.hpp"
 
-#include "app/organicmaps/platform/PlatformNew.hpp"
+#include "app/organicmaps/platform/AndroidPlatform.hpp"
 
 #include "drape/drape_diagnostics.hpp"
 #include "drape/support_manager.hpp"
